Simplify padding rules in SDG pages
We can give the padding to the whole page instead of giving it to
individual elements.

On the minus side, now the padding in the SDG pages is not the same as
the padding in the homepage, so we need to add an extra padding in only
one of these cases.